Add support for patching the compilation database
Instead of updating the whole thing.  This can be much faster, in
particular the aquery for a single target can be much faster than the
full aquery.

In setups that use other tools that should be called at the same time,
like gazelle for BUILD file generation, this can be called from a
larger script which also handles that.  I haven't tried to integrate
that into this, it would need to be a little too configurable to
integrate easily I think

This doesn't update the full dependency graph of the provided target.
In particular, updating this in the reverse direction (e.g. if modifying
the compile command for a header) would require tracking the dependency
graph used to build the current compilation database.  This would be a
good idea, but not something I'm doing here yet.

Topic: cc-patch
Reviewers: wolfd